requirements.txt
""" dependencies installed in environment """

pylint-2.3.1  # "-m pip install -U pylint"
pygame-1.9.6 # "pip install pygame"